Añadida distancia como parámetro de HotelesController. Cambiado el tipo de distancia a Double para poder utilizar valores menores a 1km.

+++ src/main/java/com/birt/berton/BertonRestAPI_PROY_1C/controllers/CasasRuralesController.java
@@ -26,14 +26,16 @@ public class CasasRuralesController {
 		return new ResponseEntity<List<CasaRural>>(casaRuralService.findAll(), HttpStatus.OK);
 	}
 	
+	//Type -> "Agroturismos"
 	@GetMapping(value="/casasrurales", params="type")
 	public ResponseEntity<List<CasaRural>> getByPropertiesLodgingtype(@RequestParam("type") String type) {
 		return new ResponseEntity<List<CasaRural>>(casaRuralService.findByPropertiesLodgingtype(type), HttpStatus.OK);
 	}
 	
+	//Type -> "Agroturismos"
 	@GetMapping(value="/casasrurales/geo", params={"lon", "lat", "dist", "type"})
 	public ResponseEntity<List<CasaRural>> getByLodgingtypeAndLocation(
-			@RequestParam("lon") Double lon, @RequestParam("lat") Double lat, @RequestParam("dist") Integer dist, @RequestParam("type") String type ) {
+			@RequestParam("lon") Double lon, @RequestParam("lat") Double lat, @RequestParam("dist") Double dist, @RequestParam("type") String type ) {
 		
 		Double radius = dist / 6378.1;
 		return new ResponseEntity<List<CasaRural>>(casaRuralService.findByLodgingtypeAndLocation(lon, lat, radius, type), HttpStatus.OK);
